multiplayer phase 1, WIP server-side Interaction, bug fixes
Added a couple small fixes (to at least prevent crashing). The game mode is always null on the client, so there's a proper check before notifying that an actor was killed.

Input binding/unbinding is now called exclusively on the client, to ensure compatibility with possess/unpossess when switching between game windows.