"CommonGame" plugin from Lyra + dependencies
This is the foundation needed for accessing a UI subsystem enabling us to push/pop menus within "layers" defined by gameplay tags. ModularGameplayActors is a somewhat odd requirement, but it's just a lightweight plugin containing abstract classes that provide consistent/easy hooks for tapping into the ModularGameplay plugin's functionality.

Could have rewritten the UI functionality myself, but figured this was a good time to get up to speed with the "Lyra" way of doing things, and getting acquainted with modular gameplay features soon can only be a good thing for prototyping.